Refactor Donna.java to improve code quality
Handle methods (handleAdd(), handleDelete(), etc.) have similar
functions.

These common functions are extracted to a new CommandHandler class.

Minor tweaks to improve code quality and readability.
